Hopefully fix the count after break issue
From testing it seems like the issue was that if there are multiple
timers running, that one of them would set the mode back to work mode
and the other one would then count one up thinking it has been in
work mode when it has been in break mode.

Somehow I figured maybe putting into one use effect might solve this
issues, but I'm not entirely sure why haha. However when testing
this seemed to solve the issue.

Anyway I also added the number of the pomodoro in the title bar.